Change CommentList component
Receiving postId prop.

Calling "useFetch" to get comments by specify post.

A null coalescing operator has been added to check if the data is
null and, if so, store the empty array in the comments variable.

Add conditional to render content.

Add conditional to checks if the item's length is 0, and if so, render
a p element.

Remove static data.