Refactored App component
Replaced UserProvider with AppProvider.

Add loading state.

Added a conditional to check if loading state is true then rendering
Loading component.

Added the "handleCheckToken' to check if token exp is expired.

Added the "handleRemoveToken" to remove local storage token item.

Refactored handleGetUser function,
a conditional is added to check whether the fetching result was
successful. If yes, then set the data. If not, then print errors
in the console.