Refactored App component
Two effect hooks have been added.

The first one is used to get the last path value from the session
storage, which is used to navigate to the path after redirecting
to the callback page after user authentication is completed.

The second checks whether the user is logged in and the
authentication has not expired. If true, get user information
and token.

Added "handleVerifyTokenExpire", "handleExChangeToken" and
handleLogout functions.