Updated index css
Removed styles for body descendant universal selector.

Added margin with 0 value for all heading and p elements.

The anchor element added color with inherit value and cursor with
pointer value.

Added margin with 0 value and  padding with 0 value for ol and ul
elements.

Removed width with 100% value for img element.

Created an img element that does not have an img element of descendant
of an id "postDetail".