Updated contact module css
Changed the padding value to "5% 10%" for class contact in the media
query for min width 450px viewport.

Changed the padding value to "5% 15%" for class contact in the media
query for min width 700px viewport.

Changed the padding value to "5% 10%" and removed padding-top for
class contact in the media query for min width 1000px viewport.