Updated PostDetail module css
Changed the row-gap value to 15px for class "postDetail".

Changed the row-gap value to 15px for class dateTime.

Created ol, ul, li, a, p, h2, h3, h5 of descendants of class
"postDetail" styles.

Created id "mainImage" of descendant of class "postDetail" styles.

Created img of descendant of id content of descendant of class
"postDetail" styles.

Created blockquote element and dark mode of descendants of class
"postDetail" styles.

Created pre element and dark mode of descendants of class
"postDetail" styles.

Changed the row-gap value to 30px for class postDetail in the
media query.

Changed the row-gap value to 40px for class postDetail in the
media query for min width 1000px viewport.

Added the padding value to "0 10%" for class imageWrap in
the media query for min width 1000px viewport.

Deleted the class title in the media query for min width
1000px viewport.