Updated App component
Added alert component and state to render fetch message.

Created a new useEffect to handle get user info.

Changed conditional for call handleGetAuthCode in the useEffect.

Added useCallback to handleTokenExpire and handleExChangeToken
for catch.

A new div with headerBar class is added to wrap the Header element.

Clean, sort and rename props and context states.