Updated Dropdown module css
Added width 70% value to dropdown class.

Added row-gap 10px value to profile class.

Created span of profile class child styles.

Added the width value to 40% for dropdown class in the media query
for min width 450px viewport.

Changed the width value to 30% for dropdown class in the media query
for min width 700px viewport.

Changed the width value to 25% for dropdown class in the media query
for min width 1000px viewport.

Changed the width value to 20% for dropdown class in the media query
for min width 1300px viewport.