feat!: updated Alert component
BREAKING CHANGE:

Check if the alert length is greater than 0 and if so,
render the alert message instead of checking if the activeAlert value is true.

Removed all "useEffect" hooks, and created startTimer and endAlert to set timer and clear timer.

Created handlePauseTimer and handleContinueTimer.

Sets the default value for all ref hooks.